Fix SHAPE_ARC hittest and adjust radius return
The hittest needs to use distance calc rather than squared distance.
This also adjusts the radius value to be double as to avoid unneeded
loss of precision

+++ libs/kimath/include/geometry/shape_arc.h
@@ -106,7 +106,7 @@ public:
 
     void Mirror( bool aX = true, bool aY = false, const VECTOR2I& aVector = { 0, 0 } );
 
-    int GetRadius() const;
+    double GetRadius() const;
 
     SEG GetChord() const
     {
